fix: refactor to built-in approach
