+++ src/shared/markdown-serializer.ts
@@ -12,6 +12,7 @@ import {
 } from "./html-helpers";
 import { IExternalPluginProvider } from "./editor-plugin";
 import MarkdownIt from "markdown-it";
+import { stackSnippetMarkdownSerializer } from "./plugins/stack-snippets/schema";
 
 // helper type so the code is a tad less messy
 export type MarkdownSerializerNodes = ConstructorParameters<
@@ -478,7 +479,7 @@ const customMarkdownSerializerNodes: MarkdownSerializerNodes = {
         state.wrapBlock(">! ", null, node, () => state.renderContent(node));
     },
 
-    //...stackSnippetSerializerNodes
+    ...stackSnippetMarkdownSerializer,
 };
 
 /**
